feat(SpaceInvaders): added moving enemies on the screen

+++ app/src/main/java/com/example/gamehub/features/spaceinvaders/classes/DataClasses.kt
@@ -2,6 +2,6 @@ package com.example.gamehub.features.spaceinvaders.classes
 
 data class Player(var x: Float, val y: Float)
 
-data class Enemy(var x: Int, var y: Int, var isAlive: Boolean = true, var type: EnemyType)
+data class Enemy(var x: Float, var y: Float, var isAlive: Boolean = true, var type: EnemyType)
 
 enum class EnemyType { SHOOTER, BOTTOM, MIDDLE }
